Updated build and add run script
Since the program no longer uses command line arguments, testing
it requires frequntly building and running it. For that reason a
simple run.sh script was added.

However, the original build.sh script is not deleted but a tad
impoved, for I may still use it, and run.sh script was made to
rely on build.sh.