Normalized types' tests to perform similar amount of operations
To do so, I found a formula for count of operations, and derived a
formula for do_one_over parameter from there.

For conviniece, I added a script.py where necessay calculations can
be performed by the python interpreter.

Also solved an issue where for loop in the main testing function
would become infinite because of uint64_t overflow.